Cannon (#42)
* Added cannon classes.

* Changed doubles to floats for consistency.

* Updated header naming.

* Added allegiance variable.

+++ src/exit.cpp
@@ -6,10 +6,10 @@
 #include "room.hpp"
 
 
-Exit::Exit(double x, double y, std::shared_ptr<Exit> destination) : Actor(x, y) {
+Exit::Exit(float x, float y, std::shared_ptr<Exit> destination) : Actor(x, y) {
 	this->destination = destination;
 }
 
-Exit::Exit(double x, double y, double width, double height, std::shared_ptr<Exit> destination) : Actor(x, y) {
+Exit::Exit(float x, float y, float width, float height, std::shared_ptr<Exit> destination) : Actor(x, y) {
 	this->destination = destination;
 }
